requirements.txt
fastapi
uvicorn
requests
feedparser
python-dotenv
supabase
resend
apscheduler
pytz
pydantic